1.优化爬虫代码 2.重构列车详情逻辑

+++ src/main/java/com/think/service/train/TrainInfoService.java
@@ -100,6 +100,8 @@ public class TrainInfoService extends ServiceImpl<ITrainInfoMapper,TrainInfoEnti
     @Transactional(rollbackFor = Exception.class)
     public void cleanDuplicateTrainInfo(){
         List<Long> tList = trainInfoMapper.selectDuplicateTrainInfo();
-        trainInfoMapper.deleteBatchIds(tList);
+        if(!tList.isEmpty()) {
+            trainInfoMapper.deleteBatchIds(tList);
+        }
     }
 }
